+++ companion/delphilint-vscode/src/delphiProjectUtils.ts
@@ -3,13 +3,15 @@ import * as fs from "fs";
 import * as path from "path";
 import { ProjectOptions } from "./projectOptions";
 
-let activeProjectFile: string | undefined = undefined;
+export type ProjectChoice = string | false;
 
-export function getActiveProject(): string | undefined {
+let activeProjectFile: ProjectChoice | undefined = undefined;
+
+export function getActiveProject(): ProjectChoice | undefined {
   return activeProjectFile;
 }
 
-function setActiveProject(value: string | undefined) {
+function setActiveProject(value: ProjectChoice) {
   activeProjectFile = value;
 }
 
@@ -19,7 +21,7 @@ async function getProjectFilesInWorkspace() {
   );
 }
 
-export async function promptProject(): Promise<string | undefined> {
+export async function promptProject(): Promise<ProjectChoice> {
   let projectFiles: string[] = await getProjectFilesInWorkspace();
 
   let quickPickItems = projectFiles
@@ -44,32 +46,40 @@ export async function promptProject(): Promise<string | undefined> {
     })
     .sort((a, b) => a.label.localeCompare(b.label));
 
+  quickPickItems.unshift({
+    label: "None (run standalone)",
+    description: "",
+    path: "__none__",
+  });
+
   let chosenItem = await vscode.window.showQuickPick(quickPickItems, {
     canPickMany: false,
     title: "DelphiLint: Choose Active Delphi Project",
     ignoreFocusOut: true,
   });
 
-  if (chosenItem) {
+  if (!chosenItem || chosenItem.path === "__none__") {
+    return false;
+  } else {
     return chosenItem.path;
   }
 }
 
-export async function promptActiveProject(): Promise<string | undefined> {
+export async function promptActiveProject(): Promise<ProjectChoice> {
   let proj = await promptProject();
-  if (proj) {
+  setActiveProject(proj);
+
+  return proj;
+}
+
+export async function getOrPromptActiveProject(): Promise<ProjectChoice> {
+  let proj: ProjectChoice | undefined = getActiveProject();
+  if (proj === undefined) {
+    proj = await promptProject();
     setActiveProject(proj);
   }
 
-  return getActiveProject();
-}
-
-export async function getOrPromptActiveProject(): Promise<string | undefined> {
-  if (!activeProjectFile) {
-    setActiveProject(await promptProject());
-  }
-
-  return getActiveProject();
+  return proj;
 }
 
 export function getProjectOptions(